Convert server to typescript
