+++ .gitignore
@@ -19,7 +19,8 @@ client/build
 .env.development.local
 .env.test.local
 .env.production.local
-/server/config/config.js
+/server/config/config.ts
+/server/dist
 
 npm-debug.log*
 yarn-debug.log*
