+++ server/config/expresss.ts
@@ -1,0 +1,43 @@
+import bodyParser from 'body-parser';
+import express from 'express';
+import mongoose from 'mongoose';
+import morgan from 'morgan';
+import path from 'path';
+import { routes } from '../routes/examples.server.routes';
+import config from './config';
+
+//     exampleRouter = require("../routes/examples.server.routes");
+
+/*
+        connect to database
+        - reference README for db uri
+    */
+mongoose.connect(process.env.DB_URI || config.db.uri, {
+  useNewUrlParser: true
+});
+mongoose.set('useCreateIndex', true);
+mongoose.set('useFindAndModify', false);
+
+// initialize app
+const app = express();
+
+// enable request logging for development debugging
+app.use(morgan('dev'));
+
+// body parsing middleware
+app.use(bodyParser.json());
+
+// add a router
+app.use('/api/example', routes);
+
+if (process.env.NODE_ENV === 'production') {
+  // Serve any static files
+  app.use(express.static(path.join(__dirname, '../../client/build')));
+
+  // Handle React routing, return all requests to React app
+  app.get('*', (req, res) => {
+    res.sendFile(path.join(__dirname, '../../client/build', 'index.html'));
+  });
+}
+
+export default app;
