--- server/controllers/examples.server.controller.js
@@ -1,5 +1,0 @@
-const Example = require('../models/examples.server.model.js')
-
-exports.hello = function(req, res) {
-    res.send('world')
-};
