+++ server/controllers/examples.server.controller.ts
@@ -1,0 +1,5 @@
+// import Example from "../models/examples.server.model";
+
+export const hello = (req: any,  res: any) => {
+    res.send('world');
+};
