+++ server/package.json
@@ -5,7 +5,7 @@
     "main": "server.js",
     "scripts": {
         "test": "mocha tests",
-        "start": "node server.js"
+        "start": "concurrently \"tsc -w\" \"nodemon dist/server.js\""
     },
     "repository": {
         "type": "git",
@@ -22,11 +22,18 @@
         "xlsx": "^0.15.1"
     },
     "devDependencies": {
+        "@types/body-parser": "^1.17.1",
+        "@types/express": "^4.17.2",
+        "@types/mailgun-js": "^0.22.3",
+        "@types/mongoose": "^5.5.30",
+        "@types/morgan": "^1.7.37",
+        "@types/node": "^12.12.7",
         "concurrently": "^4.1.0",
         "mocha": "^6.2.1",
-        "nodemon": "^1.18.10",
+        "nodemon": "^1.19.4",
         "should": "^13.2.3",
-        "supertest": "^4.0.2"
+        "supertest": "^4.0.2",
+        "typescript": "^3.7.2"
     },
     "homepage": "https://github.com/UFWebApps/React_Template_Repo#readme"
 }
