--- server/routes/examples.server.routes.js
@@ -1,8 +1,0 @@
-const examples = require('../controllers/examples.server.controller.js'),
-    express = require('express'), 
-    router = express.Router()
-
-router.route('/')
-  .get(examples.hello);
-  
-module.exports = router;
