+++ server/tslint.json
@@ -1,0 +1,12 @@
+{
+    "defaultSeverity": "error",
+    "extends": [
+        "tslint:recommended"
+    ],
+    "jsRules": {},
+    "rules": {
+        "trailing-comma": [ false ],
+        "quotemark": [ true, "single" ]
+    },
+    "rulesDirectory": []
+}
